+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # next
+* Fixed a bug in the `GEnum` instances for unsigned `Integral` types
 * Made instances polykinded where possible
 
 # 1.10.0